# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bhejna API
  description: |
    Bhejna is a high-performance WhatsApp messaging proxy and delivery engine. 
    It provides a contract-first API for sending messages, tracking delivery status, and receiving inbound messages via webhooks.
  version: 1.0.0
  contact:
    name: CodeNxt Lab
    url: https://codenxtlab.com

servers:
  - url: https://api.bhejna.com
    description: Production Server

tags:
  - name: Messages
    description: WhatsApp message delivery operations
  - name: Internal
    description: System administration and sync operations
  - name: Webhooks
    description: Inbound delivery status and message updates

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    InternalAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    MessageType:
      type: string
      enum: [text, template, image, document, audio, video, sticker, location, contacts, interactive]
      description: Allowed WhatsApp message types

    SendMessageRequest:
      type: object
      required:
        - recipient
        - message_type
        - payload
      properties:
        recipient:
          type: string
          description: "E.164 formatted phone number (e.g., +1234567890)"
          pattern: '^\+?\d{7,15}$'
          example: "+1234567890"
        message_type:
          $ref: '#/components/schemas/MessageType'
        payload:
          type: object
          description: "Meta-compatible message payload"
          example:
            body: "Hello, this is a test message from Bhejna!"

    SendMessageResponse:
      type: object
      required:
        - job_id
        - status
      properties:
        job_id:
          type: string
          example: "01ARZ3NDEKTSV4RRFFQ69G5FAV"
        status:
          type: string
          enum: [queued]

    Tenant:
      type: object
      required:
        - id
        - waba_id
        - phone_number_id
      properties:
        id:
          type: string
        waba_id:
          type: string
        phone_number_id:
          type: string
        api_key:
          type: string
        messaging_limit:
          type: integer
        quality_rating:
          type: string
        is_paused:
          type: boolean
        webhook_url:
          type: string
        webhook_secret:
          type: string
        created_at:
          type: string
          format: date-time

    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        request_id:
          type: string
          example: "req_123"

    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: "INVALID_PHONE"
            message:
              type: string
              example: "Recipient phone number invalid"
            retryable:
              type: boolean
              example: false
        request_id:
          type: string
          example: "req_123"

    WebhookPayload:
      type: object
      required:
        - object
        - entry
      properties:
        object:
          type: string
          example: "whatsapp_business_account"
        entry:
          type: array
          items:
            type: object
            required:
              - id
              - changes
            properties:
              id:
                type: string
                description: "WhatsApp Business Account ID"
              changes:
                type: array
                items:
                  type: object
                  required:
                    - field
                    - value
                  properties:
                    field:
                      type: string
                      example: "messages"
                    value:
                      type: object
                      properties:
                        messaging_product:
                          type: string
                          example: "whatsapp"
                        metadata:
                          type: object
                          properties:
                            display_phone_number:
                              type: string
                            phone_number_id:
                              type: string
                        statuses:
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: string
                              status:
                                type: string
                                enum: [sent, delivered, read, failed, deleted]
                              timestamp:
                                type: string
                              recipient_id:
                                type: string
                              errors:
                                type: array
                                items:
                                  type: object
                                  properties:
                                    code:
                                      type: integer
                                    title:
                                      type: string
                        messages:
                          type: array
                          items:
                            type: object
                            properties:
                              from:
                                type: string
                              id:
                                type: string
                              timestamp:
                                type: string
                              type:
                                type: string
                              text:
                                type: object
                                properties:
                                  body:
                                    type: string

paths:
  /v1/meta/webhook:
    get:
      operationId: GetV1MetaWebhook
      tags:
        - Webhooks
      summary: Meta Webhook Verification
      parameters:
        - in: query
          name: hub.mode
          required: true
          schema:
            type: string
        - in: query
          name: hub.challenge
          required: true
          schema:
            type: string
        - in: query
          name: hub.verify_token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
        '403':
          description: Forbidden
    post:
      operationId: PostV1MetaWebhook
      tags:
        - Webhooks
      summary: Meta Webhook Payload Receiver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookPayload'
      responses:
        '200':
          description: OK

  /v1/internal/webhook-schema:
    post:
      operationId: forceGenerateWebhookType
      description: "Internal shadow endpoint to force generation of WebhookPayload type."
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookPayload'

  /v1/messages:
    post:
      operationId: sendMessage
      tags:
        - Messages
      summary: Send a WhatsApp message
      description: Enqueues a message for delivery via WhatsApp.
      security:
        - ApiKeyAuth: []
      parameters:
        - in: header
          name: Idempotency-Key
          schema:
            type: string
          required: false
          description: "Unique key to prevent duplicate message processing"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '202':
          description: Message accepted and queued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SendMessageResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '429':
          description: Quota exceeded

  /v1/internal/tenant:
    post:
      operationId: syncTenant
      tags:
        - Internal
      summary: Sync or provision a tenant
      description: Internal endpoint to provision or update tenant configuration.
      security:
        - InternalAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/Tenant'
                - type: object
                  required:
                    - record
                  properties:
                    record:
                      $ref: '#/components/schemas/Tenant'
      responses:
        '200':
          description: Tenant synced successfully

  /v1/internal/tenants/{id}/pause:
    put:
      operationId: pauseTenant
      tags:
        - Internal
      summary: Pause a tenant
      description: Manually pause message delivery for a tenant.
      security:
        - InternalAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Tenant paused successfully

webhooks:
  delivery_update:
    post:
      operationId: handleDeliveryUpdate
      tags:
        - Webhooks
      summary: Delivery status update
      description: Sent when a message status changes (sent, delivered, read, failed).
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookPayload'
      responses:
        '200':
          description: Webhook received